Rename to unifiedpush-node-sender - AGPUSH-880
